Add files via upload
Gave a talk for IYKRA x R Ladies Jakarta collaboration (15 April 2021) and gave a talk on the topic of using SNA on Star Wars dataset. I adjusted scripts from the following two:
1. https://rpubs.com/ogimenez/200849 (github Dr. Olivier Gimenez)
2. http://pablobarbera.com/big-data-upf/html/02a-networks-intro-visualization.html
The dataset for episode 7 was transformed from .pdf to .txt, and it was a bit messy, so needed some cleaning to enable using the script made by Dr. Gimenez. The dataset for episode 4 was also a bit different from the dataset used by Dr. Gimenez, so adjusted as well. Additional edit using photoshop was also done.